backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.8,<3
groq>=1.0.0
pymongo[srv]==4.5.0
httpx>=0.28.0